+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Minor changes to README.
 
+### Fixed
+
+- Only parse exception list on startup and when it changes.
+
 ## [0.1.0] - 2024-06-22
 
 - Initial version.